Fix possible memory corruption bug
Running CouchDB on Erlang 20.3.8.5 internally has shown some very
strange behavior. Using a debug flavor beam.smp I managed to find that
the Erlang VM seems to think that there's a bug where we're clobbering
the memory fence at the end of the SnappyNifSink ErlNifBinary's buffer.

This commit is an attempt at simplifying the code to try and suss out if
it really is the Snappy NIF that's corrupting memory or if we're somehow
corrupting only after Snappy buffers from somewhere else (which would be
odd).